% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfredstats.R
\name{alfredstats}
\alias{alfredstats}
\title{alfredstats}
\usage{
alfredstats(qcdata, out)
}
\arguments{
\item{qcdata}{path to alfred qc.tsv.gz}

\item{out}{path to HTML report}
}
\value{
HTML file
}
\description{
generate an interactive HTML report with statistics from alfred qc (zgrep "^ME <qc.tsv.gz> | cut -f 2- | datamash transpose)
}
\examples{
#do not run
qcdata<-file.path("/path/to/qc.tsv.gz")
outhtml<-file.path("/path/to/qc.html")
NanoR::alfredstats(qcdata=qcdata, out=outhtml)
}
