% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_analysis.R
\name{calcDELimma}
\alias{calcDELimma}
\title{limma differential expression analysis}
\usage{
calcDELimma(nano, designfile, generatePlots = TRUE, meanCountCutoff = 5)
}
\arguments{
\item{nano}{The nano object}

\item{designfile}{A design file containin the columns file, group, include}

\item{generatePlots}{Logical whether plots should be generated as well}

\item{meanCountCutoff}{The mean count value necessary for a gene to be considered for analysis. Defaults to 5}
}
\description{
Performs differential expression analysis on the data using the limma package
}
\examples{
calcDELimma()
}
\keyword{dge}
\keyword{differential}
\keyword{expression}
\keyword{limma}

