% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_analysis.R
\name{calcFCs}
\alias{calcFCs}
\title{Fold change for all genes}
\usage{
calcFCs(counts, contrast, groups)
}
\arguments{
\item{counts}{The count matrix}

\item{contrast}{The contrast of interest}

\item{groups}{Character vectors specifying the groups of the columns}
}
\description{
Calculate fold change for all genes between a condition
}
\examples{
calcFCs()
}
\keyword{Fold}
\keyword{change}

