% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_analysis.R
\name{calcStudentPvals}
\alias{calcStudentPvals}
\title{Student t-Test}
\usage{
calcStudentPvals(counts, contrast, groups)
}
\arguments{
\item{counts}{The count matrix}

\item{contrast}{The contrast of interest}

\item{groups}{Character vectors specifying the groups of the columns}
}
\description{
Calculate P-value and t-Statistic using the t-Test
}
\examples{
calcStudentPvals()
}
\keyword{Pvalue}
\keyword{t-Test}

