% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceRatio.R
\name{distanceRatio}
\alias{distanceRatio}
\title{Distance Ratio Calculation}
\usage{
distanceRatio(mat, groups)
}
\arguments{
\item{mat}{The count matrix}

\item{groups}{The groups the samples belong to}
}
\description{
Calculates list of distance ratios for every group in a count matrix
}
\examples{
distanceRatio()
}
\keyword{distance}
\keyword{ratio}

