% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_analysis.R
\name{geneLogFC}
\alias{geneLogFC}
\title{Fold change for a gene}
\usage{
geneLogFC(g1, g2, gene, counts)
}
\arguments{
\item{g1}{Column vector of group 1}

\item{g2}{Colum vector of group 2}

\item{gene}{Integer specifying the row of the gene}

\item{counts}{The count matrix}
}
\description{
Calculate fold change of a gene between conditions
}
\examples{
geneLogFC()
}
\keyword{Fold}
\keyword{change}

