% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_analysis.R
\name{generateGroupVectors}
\alias{generateGroupVectors}
\title{Generate numeric group vectors}
\usage{
generateGroupVectors(groups, contrast)
}
\arguments{
\item{groups}{The groups the samples belong to}

\item{contrast}{The contrast for which the columns should be extracted}
}
\description{
Generate numeric vectors characterizing the group columns
}
\examples{
generateGroupVectors()
}
\keyword{column}
\keyword{group}
\keyword{vector}

