% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{geo.mean}
\alias{geo.mean}
\title{Geometric Mean}
\usage{
geo.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric data}

\item{na.rm}{Logic, if NAs should be removed or not. Defaults to TRUE}
}
\description{
Compute geometric mean
}
\examples{
geo.mean()
}
\keyword{geometric}
\keyword{mean}

