% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\title{heatmap}
\usage{
heatmap(summary, time = 1, platform = "minion", out)
}
\arguments{
\item{summary}{path to a sequencing summary file from a single ONT run}

\item{time}{hours to split the sequencing run into. Default to 1}

\item{out}{path to HTML file}

\item{platfrom}{flowcell used for sequencing (minion, promethion. Default to minion)}
}
\value{
HTML file
}
\description{
generate an interactive HTML report with channels activity over time and channels activity with respect to their disposition on the flowcell
}
\examples{
#do not run
summary<-file.path("/path/to/sequencing_summary.txt")
outhtml<-file.path("/path/to/heatmap.html")
NanoR::heatmap(summary=summary time=1,out=outhtml)
}
