% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{housekeepingCorrelation}
\alias{housekeepingCorrelation}
\title{Correlation between housekeeping genes}
\usage{
housekeepingCorrelation(nano)
}
\arguments{
\item{nano}{The nano object}
}
\description{
Plots a correlation matrix for the housekeeping genes.
}
\examples{
housekeepingCorrelation()
}
\keyword{correlation}
\keyword{housekeeping}

