% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenqual.R
\name{lenqual}
\alias{lenqual}
\title{lenqual}
\usage{
lenqual(summary, time = 1, out)
}
\arguments{
\item{summary}{path to a sequencing summary file from a single ONT run}

\item{time}{hours to split the sequencing run into. Default to 1}

\item{out}{path to HTML file}
}
\value{
HTML file
}
\description{
generate an interactive HTML report with reads length and quality over time
}
\examples{
#do not run
summary<-file.path("/path/to/sequencing_summary.txt")
outhtml<-file.path("/path/to/lenqual.html")
NanoR::lenqual(summary=summary,time=1,out=outhtml)
}
