% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lenqualdist.R
\name{lenqualdist}
\alias{lenqualdist}
\title{lenqualdist}
\usage{
lenqualdist(summary, fraction = 0.01, out)
}
\arguments{
\item{summary}{path to a sequencing summary file from a single ONT run}

\item{fraction}{fraction of reads to sample from the original sequencing summary file. Default to 0.01}

\item{out}{path to HTML file}
}
\value{
HTML file
}
\description{
generate an interactive HTML report with length and quality distributions for a subset of reads
}
\examples{
#do not run
summary<-file.path("/path/to/sequencing_summary.txt")
outhtml<-file.path("/path/to/lengthqualitydist.html")
NanoR::lenqualdist(summary=summary,fraction=.01,out=outhtml)
}
