% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxscan.R
\name{muxscan}
\alias{muxscan}
\title{muxscan}
\usage{
muxscan(muxdata, out)
}
\arguments{
\item{muxdata}{path to a mux scan data file from a single ONT run}

\item{out}{path to HTML report}
}
\value{
HTML file
}
\description{
generate an interactive HTML report with muxes status over time
}
\examples{
#do not run
muxdata<-file.path("/path/to/mux_scan_data.csv")
outhtml<-file.path("/path/to/muxscan.html")
NanoR::muxscan(muxdata=muxdata, out=outhtml)
}
