% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{nano_means}
\alias{nano_means}
\title{Combine all runs on a day / case by simple arithmetic mean}
\usage{
nano_means(df, count_var = count)
}
\arguments{
\item{df}{Dataframe holding nano data}

\item{count_vat}{column holding particle counts}
}
\description{
Combine all runs on a day / case by simple arithmetic mean
}
