% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{nsBackgroundCorrect}
\alias{nsBackgroundCorrect}
\title{Background Correction}
\usage{
nsBackgroundCorrect(nsdata, bm = "mean", sd.factor = 2)
}
\arguments{
\item{nsdata}{A nano object}

\item{bm}{Background method, currently only "mean" possible}

\item{sd.factor}{Number of standard deviations to be added to background value. Defaults to 2}
}
\description{
Calculates background noise and subtracts it from counts
}
\examples{
nsBackgroundCorrect()
}
\keyword{background}
\keyword{controls}
\keyword{correction}
\keyword{negative}

