% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{nsNormalize}
\alias{nsNormalize}
\title{Normalization functions}
\usage{
nsNormalize(nsdata, method)
}
\arguments{
\item{nsdata}{A nano object}

\item{method}{The normalization method. One of c("top100", "housekeeping", "total")}
}
\description{
Content normalization of nanostring data
}
\examples{
nsNormalize()
}
\keyword{normalization}

