% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{nsPositiveControlNormalization}
\alias{nsPositiveControlNormalization}
\title{Positive Control Normalization}
\usage{
nsPositiveControlNormalization(nsdata, pcm = "geo.mean")
}
\arguments{
\item{nsdata}{A nano object}

\item{pcm}{Positive control method, should be one of c("mean", "geo.mean")}
}
\description{
Normalize using the positive controls (spike-ins)
}
\examples{
nsPositiveControlNormalization()
}
\keyword{control}
\keyword{normalization}
\keyword{positive}

