% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rcc.R
\name{parseRCC}
\alias{parseRCC}
\title{Parse RCC files}
\usage{
parseRCC(dir = ".", rcc.pattern = ".RCC$", walkSubdirs = FALSE)
}
\arguments{
\item{dir}{The path to the RCC directory. Defaults to the current directory.}

\item{rcc.pattern}{Pattern to recognize RCC files in case the don't end with .RCC, which is assumed by default.}

\item{walkSubdirs}{If TRUE, RCC files in subdirectories will be parsed as well. Default is FALSE.}
}
\description{
Function to parse all RCC files from one directory
and merge them into a nano object
}
\examples{
parseRCC()
}
\keyword{RCC}
\keyword{parsing}

