% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_analysis.R
\name{performTAnalysis}
\alias{performTAnalysis}
\title{Student t-Test Analysis}
\usage{
performTAnalysis(counts, contrast, groups, padj.method = "BH")
}
\arguments{
\item{counts}{The count matrix}

\item{contrast}{The contrast of interest}

\item{groups}{Character vectors specifying the groups of the columns}
}
\description{
Calculate fold changes, p-values and t-statistic for a contrast
}
\examples{
performTAnalysis()
}
\keyword{Pvalue}
\keyword{change}
\keyword{fold}
\keyword{t-Statistic}
\keyword{t-Test}

