% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plotBackgrund}
\alias{plotBackgrund}
\title{Background value plot}
\usage{
plotBackgrund(nsdata)
}
\arguments{
\item{nsdata}{A nano object}
}
\description{
Plot the calculated background values
}
\examples{
plotBackground()
}
\keyword{background}
\keyword{threshold}

