% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plotBindingDensities}
\alias{plotBindingDensities}
\title{Binding Density Plot}
\usage{
plotBindingDensities(nsdata)
}
\arguments{
\item{nsdata}{A nano object}
}
\description{
Plot the binding densities for every sample for QC
}
\examples{
plotBindingDensities()
}
\keyword{binding}
\keyword{density}
\keyword{qc}

