% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{Clustered Dendrogram}
\usage{
plotDendrogram(nano, distMethod = "euclidean", clustMethod = "complete")
}
\arguments{
\item{nano}{The nano object}

\item{distMethod}{The method for distance calculation. Defaults to "euclidean".}

\item{clustMethod}{The clustering method. Defaults to "complete".}
}
\description{
Generates dendrogram for the samples
}
\examples{
plotDendrogram()
}
\keyword{cluster}
\keyword{dendrogram}

