% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceRatio.R
\name{plotDistanceRatio}
\alias{plotDistanceRatio}
\title{Distance Ratio Plot}
\usage{
plotDistanceRatio(mat, groups)
}
\arguments{
\item{mat}{The nano object}

\item{groups}{The groups the samples belong to}
}
\description{
Generates plot of distance ratio before and after normalization for each group
}
\examples{
plotDistanceRatio()
}
\keyword{distance}
\keyword{ratio}

