% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plotFOV}
\alias{plotFOV}
\title{Field of View count plot}
\usage{
plotFOV(nsdata)
}
\arguments{
\item{nsdata}{A nano object}
}
\description{
Plot the field of view count for every sample for QC
}
\examples{
plotFOV()
}
\keyword{fov}
\keyword{qc}

