% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedBoxplot.R
\name{plotGroupBoxplot}
\alias{plotGroupBoxplot}
\title{Grouped Boxplot}
\usage{
plotGroupBoxplot(nano, groups = NA, title = "Title", countCutoff = 3)
}
\arguments{
\item{nano}{The nano object}

\item{groups}{A vector defining the groups of the samples. Default: all samples belong to the same group.}

\item{countCutoff}{The mean count value a gene must have to be considered for the plot.}
}
\description{
Generate Boxplot of all expressed counts, divided into groups
}
\examples{
plotGroupBoxplot()
}
\keyword{boxplot}

