% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plotHeatmapExpression}
\alias{plotHeatmapExpression}
\title{Expression Heatmap}
\usage{
plotHeatmapExpression(nano, groups, countCutoff = 3)
}
\arguments{
\item{nano}{The nano object}

\item{countCutoff}{The mean expression cutoff to apply. Defaults to 3.}
}
\description{
Plots clustered heatmap of expression values
Low expressed genes are removed
}
\examples{
plotHeatmapExpression()
}
\keyword{expression}
\keyword{heatmap}

