% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plotHousekeepingGenes}
\alias{plotHousekeepingGenes}
\title{Housekeeping genes plot}
\usage{
plotHousekeepingGenes(nano)
}
\arguments{
\item{nano}{The nano object}
}
\description{
Plot the variation of housekeeping genes
}
\examples{
plotHousekeepingGenes()
}
\keyword{housekeeping}
\keyword{linechart}

