% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDistanceRatio.R
\name{plotNormDistanceRatio}
\alias{plotNormDistanceRatio}
\title{Distance Ratio Plot for all different normalizations}
\usage{
plotNormDistanceRatio(nano, groups, countCutoff = 5, pcm = "mean",
  bm = "geo.mean")
}
\arguments{
\item{nano}{The non-normalized nano object}

\item{groups}{The groups the samples belong to}
}
\description{
Generates plot of distance ratio before and after normalization for each group
}
\examples{
plotNormDistanceRatio()
}
\keyword{distance}
\keyword{normalization}
\keyword{ratio}

