% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plotNormFactors}
\alias{plotNormFactors}
\title{Normalization factor plot}
\usage{
plotNormFactors(nsdata)
}
\arguments{
\item{nsdata}{A nano object}
}
\description{
Plot the calculated normalization factors
}
\examples{
plotNormFactors()
}
\keyword{factor}
\keyword{normalization}

