% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plotPCA}
\alias{plotPCA}
\title{Principal Component Analysis Plots}
\usage{
plotPCA(nano, groups, countCutoff = 5)
}
\arguments{
\item{nano}{The nano object}

\item{groups}{Group information for the columns of the count matrix}

\item{countCutoff}{Mean cut value used to cut off unexpressed genes. Defaults to 5}
}
\description{
Generates PCA plots
}
\examples{
plotPCA()
}
\keyword{analysis}
\keyword{component}
\keyword{pca}
\keyword{principal}

