% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{plotPositiveScalingFactors}
\alias{plotPositiveScalingFactors}
\title{Positive scaling factor plot}
\usage{
plotPositiveScalingFactors(nsdata)
}
\arguments{
\item{nsdata}{A nano object}
}
\description{
Plot the scaling factors calculated from the positive controls
}
\examples{
plotPositiveScalingFactors()
}
\keyword{control}
\keyword{factor}
\keyword{positive}

