% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{positiveControlR2}
\alias{positiveControlR2}
\title{Plot R2 of positive control counts}
\usage{
positiveControlR2(nano)
}
\arguments{
\item{nano}{The nano object}
}
\description{
Plot correlation matrix for the top 100 genes
}
\examples{
positiveControlR2()
}
\keyword{R2}
\keyword{control}
\keyword{linear}
\keyword{positive}

