% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rcc.R
\name{readSingleRCC}
\alias{readSingleRCC}
\title{Parse single RCC file}
\usage{
readSingleRCC(path, noGenes)
}
\arguments{
\item{path}{The file path}

\item{noGenes}{The number of genes in the panel}
}
\description{
Function to parse the content of a single RCC file
}
\examples{
readSingleRCC()
}
\keyword{RCC}
\keyword{parsing}

