% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeAnnot}
\alias{removeAnnot}
\title{Remove annotation columns}
\usage{
removeAnnot(counts)
}
\arguments{
\item{counts}{Counts from a nano object}
}
\description{
Removes the annotation columns from a nano count matrix
}
\examples{
removeAnnot()
}
\keyword{utility}

