% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderReport.R
\name{renderReport}
\alias{renderReport}
\title{HTML Report Generation}
\usage{
renderReport(rcc.dir, design.file, norm.method = "housekeeping",
  pcm = "geo.mean", bm = "mean")
}
\arguments{
\item{rcc.dir}{The directory containing the RCC files}

\item{design.file}{The path to the design file}

\item{norm.method}{The method to use for RNA content normalization. Choose one of c("housekeeping", "top100", "total")}

\item{pcm}{Positive control method, the method to use for positive control normalization. Choose one of c("mean", "geo.mean")}

\item{bm}{Background method, the method to calculate the background values. Choose one of c("mean", "geo.mean")}
}
\description{
Function to generate a HTML report using Rmarkdown
}
\examples{
renderReport()
}
\keyword{HTML}
\keyword{report}

