% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{report}
\usage{
report(summary, out)
}
\arguments{
\item{summary}{path to a sequencing summary file from a single ONT run}

\item{out}{path to TSV file}
}
\value{
TSV file
}
\description{
generate a TSV file with several informations summarizing the sequencing run
}
\examples{
#do not run
summary<-file.path("/path/to/sequencing_summary.txt")
outtsv<-file.path("/path/to/summary.tsv")
NanoR::report(summary=summary,out=outtsv)
}
