% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{topGenesCorrelation}
\alias{topGenesCorrelation}
\title{Correlation Matrix Top 100 Genes}
\usage{
topGenesCorrelation(nano)
}
\arguments{
\item{nano}{The nano object}
}
\description{
Plot correlation matrix for the top 100 genes
}
\examples{
topGenesCorrelation()
}
\keyword{correlation}
\keyword{top100}

