% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(dge.result, contrast)
}
\arguments{
\item{nano}{The nano object}

\item{groups}{Group information for the columns of the count matrix}
}
\description{
Generates Volcano Plots for every contrast given
}
\examples{
plotPCA()
}
\keyword{analysis}
\keyword{component}
\keyword{pca}
\keyword{principal}

