% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield.R
\name{yield}
\alias{yield}
\title{yield}
\usage{
yield(summary, time = 1, out)
}
\arguments{
\item{summary}{path to a sequencing summary file from a single ONT run}

\item{time}{hours to split the sequencing run into. Default to 1}

\item{out}{path to HTML file}
}
\value{
HTML file
}
\description{
generate an interactive HTML report with number of reads and bases generated during the sequencing run
}
\examples{
#do not run
summary<-file.path("/path/to/sequencing_summary.txt")
outhtml<-file.path("/path/to/yield.html")
NanoR::yield(summary=summary,time=1,out=outhtml)
}
