% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(summaries, labels = NULL, time = 10, out)
}
\arguments{
\item{summaries}{a character vector containing paths to multiple sequencing summaries (each from a different ONT sequencing run)}

\item{labels}{a character vector containing ordered names for the sequencing files. Override the default behaviour, which is to use the run_id}

\item{time}{compare run statistics using this time interval (hours). Default to 10}

\item{out}{path to HTML report}
}
\value{
HTML file
}
\description{
generate an interactive HTML report comparing length and quality of reads from multiple sequencing summary files
}
\examples{
#do not run
exp1<-file.path("/path/to/exp1_sequencing_summary.txt")
exp2<-file.path("/path/to/exp2_sequencing_summary.txt")
outhtml<-file.path("/path/to/comparison.html")
NanoR::compare(summaries=c(exp1,exp2), time=10, out=outhtml)
}
