% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{nano_means}
\alias{nano_means}
\title{Combine all runs on a day / case by simple arithmetic mean}
\usage{
nano_means(df, mean_var = count, ...)
}
\arguments{
\item{df}{Dataframe holding nano data}

\item{mean_var}{column holding to aggregate to mean}

\item{...}{variables to aggregate on}
}
\description{
Combine all runs on a day / case by simple arithmetic mean
}
